리뷰 반영
- 기존 YdsScaffold 대체
- preview 맨 아래로, private 변경
- 기존 다른 파일들도 preview function은 전부 private하게 변경
- 인터페이스, enum 맨 위로
- YdsEasing 사용
- Toast Animation 파일 분리
- 기타 코드 정리
