+++ compose/src/main/java/com/yourssu/design/system/compose/atom/Toggle.kt
@@ -87,7 +87,7 @@ fun Toggle(
 
 @Preview(showBackground = true)
 @Composable
-fun TogglePreview() {
+private fun TogglePreview() {
     var checked1 by remember { mutableStateOf(false) }
     var disabled by remember { mutableStateOf(false) }
     var checked2 by remember { mutableStateOf(false) }
