+++ compose/src/main/java/com/yourssu/design/system/compose/foundation/ToastAnimation.kt
@@ -1,0 +1,105 @@
+package com.yourssu.design.system.compose.foundation
+
+import androidx.compose.animation.core.Animatable
+import androidx.compose.animation.core.AnimationSpec
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.layout.Box
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.RecomposeScope
+import androidx.compose.runtime.State
+import androidx.compose.runtime.currentRecomposeScope
+import androidx.compose.runtime.key
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.graphicsLayer
+import com.yourssu.design.system.compose.rule.Duration
+import com.yourssu.design.system.compose.rule.YdsInAndOutEasing
+
+typealias FadeInFadeOutTransition = @Composable (content: @Composable () -> Unit) -> Unit
+
+class FadeInFadeOutState<T> {
+    // we use Any here as something which will not be equals to the real initial value
+    var current: Any? = Any()
+    var items = mutableListOf<FadeInFadeOutAnimationItem<T>>()
+    var scope: RecomposeScope? = null
+}
+
+data class FadeInFadeOutAnimationItem<T>(
+    val key: T,
+    val transition: FadeInFadeOutTransition
+)
+
+@Composable
+fun FadeInFadeOutWithScale(
+    current: ToastData?,
+    modifier: Modifier = Modifier,
+    content: @Composable (ToastData) -> Unit
+) {
+    val state = remember { FadeInFadeOutState<ToastData?>() }
+    if (current != state.current) {
+        state.current = current
+        val keys = state.items.map { it.key }.toMutableList()
+        if (!keys.contains(current)) {
+            keys.add(current)
+        }
+        state.items.clear()
+        keys.filterNotNull().mapTo(state.items) { key ->
+            FadeInFadeOutAnimationItem(key) { children ->
+                val isVisible = key == current
+                val duration = Duration.Medium.millis
+                val opacity = animatedOpacity(
+                    animation = tween(
+                        easing = YdsInAndOutEasing,
+                        delayMillis = 0,
+                        durationMillis = duration
+                    ),
+                    visible = isVisible,
+                    onAnimationFinish = {
+                        if (key != state.current) {
+                            // leave only the current in the list
+                            state.items.removeAll { it.key == key }
+                            state.scope?.invalidate()
+                        }
+                    }
+                )
+                Box(
+                    Modifier
+                        .graphicsLayer(
+                            alpha = opacity.value
+                        )
+                ) {
+                    children()
+                }
+            }
+        }
+    }
+    Box(modifier) {
+        state.scope = currentRecomposeScope
+        state.items.forEach { (item, opacity) ->
+            key(item) {
+                opacity {
+                    content(item!!)
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun animatedOpacity(
+    animation: AnimationSpec<Float>,
+    visible: Boolean,
+    onAnimationFinish: () -> Unit = {}
+): State<Float> {
+    val alpha = remember { Animatable(if (!visible) 1f else 0f) }
+    LaunchedEffect(visible) {
+        alpha.animateTo(
+            if (visible) 1f else 0f,
+            animationSpec = animation
+        )
+        onAnimationFinish()
+    }
+    return alpha.asState()
+}
+
