+++ compose/src/main/java/com/yourssu/design/system/compose/foundation/ToastHost.kt
@@ -1,23 +1,12 @@
 package com.yourssu.design.system.compose.foundation
 
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.AnimationSpec
-import androidx.compose.animation.core.CubicBezierEasing
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.RecomposeScope
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.State
-import androidx.compose.runtime.currentRecomposeScope
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.graphicsLayer
 import com.yourssu.design.system.compose.component.Toast
 import kotlinx.coroutines.CancellableContinuation
 import kotlinx.coroutines.delay
@@ -25,6 +14,23 @@ import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlin.coroutines.resume
+
+
+
+interface ToastData {
+    val message: String
+    val duration: ToastDuration
+
+    fun dismiss()
+}
+
+enum class ToastResult {
+    Dismissed,
+}
+
+enum class ToastDuration {
+    SHORT, LONG,
+}
 
 
 @Stable
@@ -46,22 +52,30 @@ class ToastHostState {
             currentToastData = null
         }
     }
+}
 
-    @Stable
-    private class ToastDataImpl(
-        override val message: String,
-        override val duration: ToastDuration,
-        private val continuation: CancellableContinuation<ToastResult>
-    ) : ToastData {
+@Stable
+private class ToastDataImpl(
+    override val message: String,
+    override val duration: ToastDuration,
+    private val continuation: CancellableContinuation<ToastResult>
+) : ToastData {
 
-        override fun dismiss() {
-            if (continuation.isActive) {
-                continuation.resume(ToastResult.Dismissed)
-            }
+    override fun dismiss() {
+        if (continuation.isActive) {
+            continuation.resume(ToastResult.Dismissed)
         }
     }
 }
 
+
+/**
+ *  Toast를 보여주기 위해선 ToastHost를 사용해야 합니다.
+ * ex) toastHostState.showToast(
+ *                    message = "Snackbar",
+ *                    duration = ToastDuration.SHORT
+ *                  )
+ */
 @Composable
 fun ToastHost(
     toastHostState: ToastHostState,
@@ -72,8 +86,8 @@ fun ToastHost(
     LaunchedEffect(currentToastData) {
         if (currentToastData != null) {
             when (currentToastData.duration) {
-                ToastDuration.SHORT -> delay(1500)
-                ToastDuration.LONG -> delay(3000)
+                ToastDuration.SHORT -> delay(ToastDurationShort)
+                ToastDuration.LONG -> delay(ToastDurationLong)
             }
             currentToastData.dismiss()
         }
@@ -85,121 +99,5 @@ fun ToastHost(
     )
 }
 
-
-interface ToastData {
-    val message: String
-    val duration: ToastDuration
-
-    fun dismiss()
-}
-
-enum class ToastResult {
-    Dismissed,
-}
-
-enum class ToastDuration {
-    SHORT, LONG,
-}
-
-
-@Composable
-private fun FadeInFadeOutWithScale(
-    current: ToastData?,
-    modifier: Modifier = Modifier,
-    content: @Composable (ToastData) -> Unit
-) {
-    val state = remember { FadeInFadeOutState<ToastData?>() }
-    if (current != state.current) {
-        state.current = current
-        val keys = state.items.map { it.key }.toMutableList()
-        if (!keys.contains(current)) {
-            keys.add(current)
-        }
-        state.items.clear()
-        keys.filterNotNull().mapTo(state.items) { key ->
-            FadeInFadeOutAnimationItem(key) { children ->
-                val isVisible = key == current
-                val duration = if (isVisible) ToastFadeInMillis else ToastFadeOutMillis
-                val opacity = animatedOpacity(
-                    animation = tween(
-                        easing = CubicBezierEasing(0.25f, 0.1f, 0.25f, 1f),
-                        delayMillis = 0,
-                        durationMillis = duration
-                    ),
-                    visible = isVisible,
-                    onAnimationFinish = {
-                        if (key != state.current) {
-                            // leave only the current in the list
-                            state.items.removeAll { it.key == key }
-                            state.scope?.invalidate()
-                        }
-                    }
-                )
-                Box(
-                    Modifier
-                        .graphicsLayer(
-                            alpha = opacity.value
-                        )
-                ) {
-                    children()
-                }
-            }
-        }
-    }
-    Box(modifier) {
-        state.scope = currentRecomposeScope
-        state.items.forEach { (item, opacity) ->
-            key(item) {
-                opacity {
-                    content(item!!)
-                }
-            }
-        }
-    }
-}
-
-private class FadeInFadeOutState<T> {
-    // we use Any here as something which will not be equals to the real initial value
-    var current: Any? = Any()
-    var items = mutableListOf<FadeInFadeOutAnimationItem<T>>()
-    var scope: RecomposeScope? = null
-}
-
-private data class FadeInFadeOutAnimationItem<T>(
-    val key: T,
-    val transition: FadeInFadeOutTransition
-)
-
-private typealias FadeInFadeOutTransition = @Composable (content: @Composable () -> Unit) -> Unit
-
-@Composable
-private fun animatedOpacity(
-    animation: AnimationSpec<Float>,
-    visible: Boolean,
-    onAnimationFinish: () -> Unit = {}
-): State<Float> {
-    val alpha = remember { Animatable(if (!visible) 1f else 0f) }
-    LaunchedEffect(visible) {
-        alpha.animateTo(
-            if (visible) 1f else 0f,
-            animationSpec = animation
-        )
-        onAnimationFinish()
-    }
-    return alpha.asState()
-}
-
-@Composable
-private fun animatedScale(animation: AnimationSpec<Float>, visible: Boolean): State<Float> {
-    val scale = remember { Animatable(if (!visible) 1f else 0.8f) }
-    LaunchedEffect(visible) {
-        scale.animateTo(
-            if (visible) 1f else 0.8f,
-            animationSpec = animation
-        )
-    }
-    return scale.asState()
-}
-
-private const val ToastFadeInMillis = 250
-private const val ToastFadeOutMillis = 250
+private const val ToastDurationShort = 1500L
+private const val ToastDurationLong = 3000L
